% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{apply_standard_emotions}
\alias{apply_standard_emotions}
\title{apply standard emotions to data.}
\usage{
apply_standard_emotions(data, remove_words)
}
\arguments{
\item{data}{\itemize{
\item A data frame that must contain the "FragmentID" column and "fragment" column (with this spelling/case) containing the text.
}}

\item{remove_words}{\itemize{
\item a vector of words to remove in addition to the standard stopwords for the language. ß
}}
}
\value{
A data frame with the sentiments and counts to plot.
}
\description{
This function returns the standard "Very Positive", "Positive", "Neutral", "Negative", "Very Negative" sentiments based on the passed in text and stop words.
}
