% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{apply_standard_emotions_german}
\alias{apply_standard_emotions_german}
\title{apply standard emotions to german text data.}
\usage{
apply_standard_emotions_german(data, remove_words)
}
\arguments{
\item{data}{\itemize{
\item A data frame that must contain the "FragmentID" column and "fragment" column (with this spelling/case) containing the text.
}}

\item{remove_words}{\itemize{
\item stop words to remove in addition to the quanteda stop words.
}}
}
\value{
A data frame with the sentiments and counts to plot.
}
\description{
This function returns the standard "Positive","Negative", sentiments based on the passed in text and stop words for German language.
}
