% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{are_rows_zero}
\alias{are_rows_zero}
\title{Which rows of a numeric data frame are all zeros}
\usage{
are_rows_zero(dta)
}
\arguments{
\item{dta}{\itemize{
\item The data that can be turned into a numeric matrix, such as a data frame or term frequency matrix
}}
}
\value{
Returns a vector of truth values, TRUE if the whole row is zero otherwise FALSE.
}
\description{
This function takes in an object that can be turned into a numeric matrix and returns for each row, whether they are all zero.
}
