% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{areColors}
\alias{areColors}
\title{Is the colour code or name valid.}
\usage{
areColors(x)
}
\arguments{
\item{x}{A vector of colour codes/characters to test}
}
\description{
Determines whether colour characters are valid colour codes/names
}
