% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{build_pair_datasets}
\alias{build_pair_datasets}
\title{Build a set of pairs data frames - one data frame binding rows from two data frames.}
\usage{
build_pair_datasets(
  framework_data,
  pairs_definitions,
  doc_var = "none",
  plot_col = "none"
)
}
\arguments{
\item{framework_data}{\itemize{
\item The framework data object.
}}

\item{pairs_definitions}{\itemize{
\item A data frame that must have at least two columns, one with name "from_id" the other "to_id" which are both data frame names in framework_data$data
}}

\item{doc_var}{\itemize{
\item Default "none", indicating the document variable name to give each dataset being paired. This will add a new column to the resultant data frame whose values indicate which data frame it came from. Used in many stats functions like keyness indicators. "none" means leave the data frame as per the originals. "auto" means use the from and to ids as the doc variable. Otherwise a data frame with from_var and to_var column names with the same number of rows as the pairs_definitions data frame.
}}

\item{plot_col}{\itemize{
\item Default "none", indicating the document variable colours to use in any pair graphing. This will add a new column to the resultant data frame whose values indicate which colour to use. Used in many stats functions like keyness indicators. "none" means leave the data frame as per the originals. "auto" means use the from and to colours as the doc colours Otherwise a data frame with from_col and to_col column names with the same number of rows as the pairs_definitions data frame.
}}
}
\value{
Returns a named list of tokens_stem (the stemmed tokens with stop words removed), tokens_unstem (the unstemmed tokens with stop words removed), dtm (the document term matrix not trimmed) and dtm_trim, the document term matrix trimmed to the min_term_freq value.
}
\description{
This function takes a dataframe of pairs definitions from existing data frames and creates new paired data frames.
}
