% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signifier_helpers.R
\name{calculate_dyad_zone}
\alias{calculate_dyad_zone}
\title{get a dyad value's zone or stones x or y zone}
\usage{
calculate_dyad_zone(left, right = 100 - left)
}
\arguments{
\item{left}{percentage triad left value (between 0 and 100).}

\item{right}{percentage triad right value (between 0 and 100). default 100 = left (enabling the pass of the left only)}
}
\value{
A single character returning the zone
}
\description{
get a dyad value's zone or stones x or y zone
}
\examples{
l <- 65
r <- 35
calculate_dyad_zone(l, r)
}
