% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{calculate_multi_select_correlations}
\alias{calculate_multi_select_correlations}
\title{do a correlation test on multi-select lists (mcqs)}
\usage{
calculate_multi_select_correlations(
  correlation_pairs,
  fwd,
  list_ids = NULL,
  test_type = "Pearson"
)
}
\arguments{
\item{correlation_pairs}{\itemize{
\item The pairs of data filters to perform tests on. This can be either the file name or the data frame created from it or within the calling code. It must have columns at least "from_id", "to_id" where the ids are the names of the data frames held in the data list in the sensemakerdatar object.
}}

\item{fwd}{\itemize{
\item The sensemakerdatar object for the framework being processed.
}}

\item{list_ids}{\itemize{
\item Default NULL otherwise the signifier ids of the multi-select lists to be processed. If NULL then all the multiselect lists will be processed.
}}

\item{test_type}{\itemize{
\item default Pearson", but the only type currently supported.
}}
}
\value{
Returns a list, the outer being one entry per correlation pair, the second being each multi-select list id, the third being each column in the dataset, and this list contains the test result list. This list has entries names "data", "expected", "residuals", "residuals_sqr", "p-value", "test_result", "used_p_simulation".
}
\description{
This function will perform a correlation test (Pearson's) for passed in list ids (or all if NULL) and a set list of correlation pair filters based on filtered cohort definitions.
}
