% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signifier_helpers.R
\name{calculate_stone_4_zone}
\alias{calculate_stone_4_zone}
\title{get a stone value's 4 zone}
\usage{
calculate_stone_4_zone(x, y)
}
\arguments{
\item{x}{fraction stone x value.}

\item{y}{fraction stone y value.}
}
\value{
A character string returning the zone
}
\description{
get a stone value's 4 zone
}
\examples{
l <- 0.7
t <- 0.3
calculate_stone_4_zone(l, t)
}
