% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signifier_helpers.R
\name{calculate_stone_x_y_zone}
\alias{calculate_stone_x_y_zone}
\title{get a stone value's x or y zone}
\usage{
calculate_stone_x_y_zone(x)
}
\arguments{
\item{x}{fraction stone x or y left value (between 0 and 1).}
}
\value{
A single character returning the zone
}
\description{
get a stone value's x or y zone
}
\examples{
l <- 0.5
calculate_stone_x_y_zone(l)
}
