% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{clearAll}
\alias{clearAll}
\title{remove all of the characters  from a string}
\usage{
clearAll(input_string, double_space_to_single = TRUE)
}
\arguments{
\item{input_string}{\itemize{
\item the string to remove all the tags.
}}

\item{double_space_to_single}{\itemize{
\item Default TRUE, will replace any remaining double space with a single space.
}}
}
\value{
The string with brackets removed.
}
\description{
remove HTML, dash, brackets, close dots, commas, quotes, points, semicolons, forward slashes, apostrophes, ampersands, percentages, long dashes, plus and colons.
}
