% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{clearAmps}
\alias{clearAmps}
\title{remove ampersands  from a string}
\usage{
clearAmps(input_string, double_space_to_single = TRUE)
}
\arguments{
\item{input_string}{\itemize{
\item the string to remove ampersands.
}}

\item{double_space_to_single}{\itemize{
\item Default TRUE, will replace any remaining double space with a single space.
}}
}
\value{
The string with ampersands removed.
}
\description{
remove ampersands  from a string
}
