% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{clearApos}
\alias{clearApos}
\title{remove apostrophes  from a string}
\usage{
clearApos(input_string, double_space_to_single = TRUE)
}
\arguments{
\item{input_string}{\itemize{
\item the string to remove apostrophes.
}}

\item{double_space_to_single}{\itemize{
\item Default TRUE, will replace any remaining double space with a single space.
}}
}
\value{
The string with apostrophes removed.
}
\description{
remove apolstrophes  from a string
}
