% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{clearBrackets}
\alias{clearBrackets}
\title{remove brackets  from a string}
\usage{
clearBrackets(input_string, double_space_to_single = TRUE)
}
\arguments{
\item{input_string}{\itemize{
\item the string to remove brackets.
}}

\item{double_space_to_single}{\itemize{
\item Default TRUE, will replace any remaining double space with a single space.
}}
}
\value{
The string with brackets removed.
}
\description{
remove brackets  from a string
}
