% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{clearCloseDots}
\alias{clearCloseDots}
\title{remove close triple dots  from a string}
\usage{
clearCloseDots(input_string, double_space_to_single = TRUE)
}
\arguments{
\item{input_string}{\itemize{
\item the string to remove close triple dots.
}}

\item{double_space_to_single}{\itemize{
\item Default TRUE, will replace any remaining double space with a single space.
}}
}
\value{
The string with close triple dots removed.
}
\description{
remove close triple dots  from a string
}
