% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{clearColon}
\alias{clearColon}
\title{remove colons  from a string}
\usage{
clearColon(input_string, double_space_to_single = TRUE)
}
\arguments{
\item{input_string}{\itemize{
\item the string to remove colons.
}}

\item{double_space_to_single}{\itemize{
\item Default TRUE, will replace any remaining double space with a single space.
}}
}
\value{
The string with colons removed.
}
\description{
remove colons  from a string
}
