% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{clearComma}
\alias{clearComma}
\title{remove commas  from a string}
\usage{
clearComma(input_string, double_space_to_single = TRUE)
}
\arguments{
\item{input_string}{\itemize{
\item the string to remove commas.
}}

\item{double_space_to_single}{\itemize{
\item Default TRUE, will replace any remaining double space with a single space.
}}
}
\value{
The string with commas removed.
}
\description{
remove commas  from a string
}
