% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{clearDash}
\alias{clearDash}
\title{remove dashes from a string}
\usage{
clearDash(input_string, double_space_to_single = TRUE)
}
\arguments{
\item{input_string}{\itemize{
\item the string to remove dash.
}}

\item{double_space_to_single}{\itemize{
\item Default TRUE, will replace any remaining double space with a single space.
}}
}
\value{
The string with dash removed.
}
\description{
Remove dashes from a string
}
