% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{clearLongDash}
\alias{clearLongDash}
\title{remove long dashes from a string}
\usage{
clearLongDash(input_string, double_space_to_single = TRUE)
}
\arguments{
\item{input_string}{\itemize{
\item the string to remove long dash.
}}

\item{double_space_to_single}{\itemize{
\item Default TRUE, will replace any remaining double space with a single space.
}}
}
\value{
The string with long dash removed.
}
\description{
Remove long dashes from a string
}
