% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{clearPer}
\alias{clearPer}
\title{remove percentages  from a string}
\usage{
clearPer(input_string, double_space_to_single = TRUE)
}
\arguments{
\item{input_string}{\itemize{
\item the string to remove percentages.
}}

\item{double_space_to_single}{\itemize{
\item Default TRUE, will replace any remaining double space with a single space.
}}
}
\value{
The string with percentages removed.
}
\description{
remove percentages  from a string
}
