% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{clearPlus}
\alias{clearPlus}
\title{remove plus sign from a string}
\usage{
clearPlus(input_string, double_space_to_single = TRUE)
}
\arguments{
\item{input_string}{\itemize{
\item the string to remove plus sign.
}}

\item{double_space_to_single}{\itemize{
\item Default TRUE, will replace any remaining double space with a single space.
}}
}
\value{
The string with plus sign removed.
}
\description{
remove plus sign from a string
}
