% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{clearPoint}
\alias{clearPoint}
\title{remove points  from a string}
\usage{
clearPoint(input_string, double_space_to_single = TRUE)
}
\arguments{
\item{input_string}{\itemize{
\item the string to remove points.
}}

\item{double_space_to_single}{\itemize{
\item Default TRUE, will replace any remaining double space with a single space.
}}
}
\value{
The string with points removed.
}
\description{
remove points  from a string
}
