% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{clearQuote}
\alias{clearQuote}
\title{remove single quotes  from a string}
\usage{
clearQuote(input_string, double_space_to_single = TRUE)
}
\arguments{
\item{input_string}{\itemize{
\item the string to remove single quotes.
}}

\item{double_space_to_single}{\itemize{
\item Default TRUE, will replace any remaining double space with a single space.
}}
}
\value{
The string with single quotes removed.
}
\description{
remove single quotes  from a string
}
