% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{clearSemiColon}
\alias{clearSemiColon}
\title{remove semi colons  from a string}
\usage{
clearSemiColon(input_string, double_space_to_single = TRUE)
}
\arguments{
\item{input_string}{\itemize{
\item the string to remove semicolon.
}}

\item{double_space_to_single}{\itemize{
\item Default TRUE, will replace any remaining double space with a single space.
}}
}
\value{
The string with semi colons removed.
}
\description{
remove semicolons  from a string
}
