% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{clearSlash}
\alias{clearSlash}
\title{remove forward slashes  from a string}
\usage{
clearSlash(input_string, double_space_to_single = TRUE)
}
\arguments{
\item{input_string}{\itemize{
\item the string to remove forward slashes.
}}

\item{double_space_to_single}{\itemize{
\item Default TRUE, will replace any remaining double space with a single space.
}}
}
\value{
The string with forward slashes removed.
}
\description{
remove forward slashes  from a string
}
