% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{clearSpace}
\alias{clearSpace}
\title{remove spaces  from a string}
\usage{
clearSpace(input_string, double_space_to_single = TRUE)
}
\arguments{
\item{input_string}{\itemize{
\item the string to remove spaces.
}}

\item{double_space_to_single}{\itemize{
\item Default TRUE, will replace any remaining double space with a single space.
}}
}
\value{
The string with spaces removed.
}
\description{
remove spaces  from a string
}
