% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{debug_log}
\alias{debug_log}
\title{Log debug information}
\usage{
debug_log(..., file = "debug_log.txt")
}
\arguments{
\item{...}{A comma separated list of parameters named by the programmer to pass data into the log.}

\item{file}{\itemize{
\item default "debug_log.txt" File name to use - this can be altered in a loop if a single file for each loop iteration is required otherwise values are appended.
}}

\item{replacement_value}{\itemize{
\item A string to use as a replacement in the wrap break - i.e. instead of breaking.
}}
}
\value{
NULL - outputs a file.
}
\description{
This helper function is very useful when executing helper functions (or any other) from within an rmd file as normal cat/print do not output unless the document output is successful.
It logs a set of properites passed and out to a file named in the out file (otherwise it is appended to the last file)
}
