% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{do_means_tests}
\alias{do_means_tests}
\title{Do means tests}
\usage{
do_means_tests(
  means_tests,
  fwd,
  signifier_ids = NULL,
  signifier_types = NULL,
  test_type = "hotelling",
  non_parametric = TRUE,
  b_value = 1000,
  keep_only_include = TRUE
)
}
\arguments{
\item{means_tests}{\itemize{
\item means tests from a means test file. Can be either the file name or the data frame created from it. It must have columns at least "from_id", "to_id" and "name". The from_id and to_id are names of data filters in the filter parameter.
}}

\item{fwd}{\itemize{
\item The sensemakerdatar object for the framework being processed.
}}

\item{signifier_ids}{\itemize{
\item Default NULL otherwise the signifier ids of the signifiers to be processed. These should be either triad and/or dyad signifier ids. If NULL then the signifier types parameter must be entered.
}}

\item{signifier_types}{\itemize{
\item Default NULL otherwise the signifier types to be processed. Values should be either "triad", "dyad" or both. If NULL then the signifier ids parameter must be entered.
}}

\item{test_type}{\itemize{
\item default "hotelling". Must be one of the supported test types, which can be found by calling sensemakerhelpersr::get_avalable_tests()
}}

\item{non_parametric}{\itemize{
\item default TRUE, otherwise FALSE. Only applicable (for now) in the hotelling test, the other supported tests are all non-parametric.
}}

\item{b_value}{\itemize{
\item the bootstrap iteration value for many of the non-parametric tests. Default 1000.
}}

\item{keep_only_include}{\itemize{
\item Default TRUE, keep only those signifiers that have the keep_only set to TRUE otherwise retrieve all signifiers.
}}
}
\value{
Returns a list, one for each signifier id, of the test results in the standard CE SenseMaker Workbench format.
}
\description{
Do means tests on the data passed into the function.
}
