% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{enough_rows}
\alias{enough_rows}
\title{Are there sufficient rows to produce a keyness graph}
\usage{
enough_rows(dta)
}
\arguments{
\item{dta}{\itemize{
\item The data that can be turned into a numeric matrix, such as a data frame or term frequency matrix
}}
}
\value{
TRUE or FALSE, the term document matrix can be printed
}
\description{
This function takes in an object that can be turned into a numeric matrix and returns TRUE or FALSE as to whether there is sufficient rows in a term document matrix to produce a keyness plot.
}
