% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{get_avalable_tests}
\alias{get_avalable_tests}
\title{Return the statistical output from a Hotelling T test.}
\usage{
get_avalable_tests()
}
\description{
A list of the available colour brewer palettes available for heat maps.
}
