% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signifier_helpers.R
\name{get_caption_values}
\alias{get_caption_values}
\title{Get the graph caption values}
\usage{
get_caption_values(filtered_data, full_data, sig_id, framework_object)
}
\arguments{
\item{filtered_data}{\itemize{
\item the data being graphed
}}

\item{full_data}{\itemize{
\item the full capture data set
}}

\item{sig_id}{\itemize{
\item the signifier id being plotted.
}}

\item{framework_object, }{the sensemakerframeworkr object for the capture.}
}
\value{
A string containing the caption values.
}
\description{
Get a string containing the caption values from the passsed in dataset. This includes data count, N/A count, not responded count and so on.
}
