% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{get_colour_palette_names}
\alias{get_colour_palette_names}
\title{Return the RColorBrewer palette names usable for heatmaps.}
\usage{
get_colour_palette_names()
}
\value{
A character vector of available colour palette names.
}
\description{
A list of the available colour brewer palettes available for heat maps.
}
