% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{get_correlations_by_type}
\alias{get_correlations_by_type}
\title{Get the correlation data for signifiers of a from and to signifier type}
\usage{
get_correlations_by_type(
  df,
  fw,
  from_type,
  to_type,
  from_signifier_classes = "signifier",
  to_signifier_classes = "signifier",
  from_signifier_ids = NULL,
  to_signifier_ids = NULL,
  keep_only_include = TRUE,
  round_digits = 0,
  residual_threshold = 4,
  p_threshold = 0.05
)
}
\arguments{
\item{df}{\itemize{
\item The data to perform the tests - this is a data frame containing at least all the columns needed for the signifier types passed in. .
}}

\item{fw}{\itemize{
\item The framework definition object.
}}

\item{from_type}{\itemize{
\item Either "list" or one of the shape types ("dyad", "triad", "stones").
}}

\item{to_type}{\itemize{
\item Either "list" or one of the shape types ("dyad", "triad", "stones").
}}

\item{from_signifier_classes}{\itemize{
\item default "signifier" - a vector of classes of signifiers to include in the from type (e.g. for a "list" one may select "date")
}}

\item{to_signifier_classes}{\itemize{
\item default "signifier" - a vector of classes of signifiers to include in the to type (e.g. for a "list" one may select "date")
}}

\item{from_signifier_ids}{\itemize{
\item default NULL - further qualify the from signifiers to include by providing a vector of signifier ids (of type "from_type")
}}

\item{to_signifier_ids}{\itemize{
\item default NULL - further qualify the to signifiers to include by providing a vector of signifier ids (of type "to_type")
}}

\item{keep_only_include}{\itemize{
\item default TRUE, to exclude any signifier that has include FALSE.
}}

\item{round_digits}{\itemize{
\item default 0, z, z^2, expected value table rounding digits.
}}

\item{residual_threshold}{\itemize{
\item default 4, the threshold for residual calculation on residual squared values.
}}

\item{p_threshold}{\itemize{
\item default 0.05, the p-value threshold for null hypothesis test.
}}
}
\value{
Returns a named list. ids_to_output contains the dataframe of signifier id pairs included in the output. sig_residuals a list of length the number of correlation pairs, It contains the "data", "expected", "residuals", "residuals_sqr", "p-value".
}
\description{
This function will perform goodness of fit tests and return the results for each signifier combination of the from and to signifier types. For lists, these will be only the single select signifier mcqs and the date columns.
The actual tests are performed on the zone and region categorical columns for triads, dyads and stones.
}
