% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signifier_helpers.R
\name{get_graph_title_size}
\alias{get_graph_title_size}
\title{Get a size value for plotting the graph titles}
\usage{
get_graph_title_size(title, title_size_multiplier = 1)
}
\arguments{
\item{title}{\itemize{
\item A string containing the title.
}}

\item{title_size_multiplier.}{A multiplier to expand or contract the retured size.}
}
\value{
An integer for the title plot size.
}
\description{
This might be a useful function - it is used by the workbench to size plot titles and useful for Workbench sized plot objects.
}
