% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{get_iso_codes}
\alias{get_iso_codes}
\title{Get the supported iso codes and place in a data frame.}
\usage{
get_iso_codes()
}
\value{
The iso codes in a data frame
}
\description{
Get iso codes supported in the application (particularly for word stemming)
}
