% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{get_palette_colour_count}
\alias{get_palette_colour_count}
\title{Return the colour count (number of distinct colours) available from a given RColorBrewer palette.}
\usage{
get_palette_colour_count(brew_colour_name)
}
\arguments{
\item{brew_colour_name}{\itemize{
\item the colour palette name
}}
}
\value{
An integer with the number of palette colours.
}
\description{
Get the number of colours a colour palette contains to help sort bin counts in heatmaps.
}
