% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{get_RColorBrewer_pal_names}
\alias{get_RColorBrewer_pal_names}
\title{Return the RColorBrewer palette names usable for discrete colouring.}
\usage{
get_RColorBrewer_pal_names()
}
\value{
A character vector of available colour palette names.
}
\description{
A list of the available colour brewer palettes available for discrete plot colours.
}
