% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{get_stones_region_correlations}
\alias{get_stones_region_correlations}
\title{Return correlation data for list class "region" for a given stone id.}
\usage{
get_stones_region_correlations(
  data_filter,
  df,
  fw,
  stones_id,
  stone_ids = NULL,
  round_digits = 4,
  residual_threshold = 4,
  p_threshold = 0.05
)
}
\arguments{
\item{df}{\itemize{
\item The data to perform the tests - this is a data frame containing at least all the columns needed for the signifier types passed in. .
}}

\item{fw}{\itemize{
\item The framework definition object.
}}

\item{stones_id}{\itemize{
\item The id for a valid framework stone.
}}

\item{stone_ids}{\itemize{
\item Default NULL, optional, a list of at least two stone ids from the region list ids to restrict the stones included in the output.
}}

\item{round_digits}{\itemize{
\item default 0, z, z^2, expected value table rounding digits.
}}

\item{residual_threshold}{\itemize{
\item default 4, the threshold for residual calculation on residual squared values.
}}

\item{p_threshold}{\itemize{
\item default 0.05, the p-value threshold for null hypothesis test.
}}
}
\value{
Returns a named list. ids_to_output contains the dataframe of signifier id pairs included in the output. sig_residuals a list of length the number of correlation pairs, It contains the "data", "expected", "residuals", "residuals_sqr", "p-value".
}
\description{
This function will perform goodness of fit tests and return the results for a given stone regions for each stone pointer.
}
