% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signifier_helpers.R
\name{get_triad_anchor_plot_size}
\alias{get_triad_anchor_plot_size}
\title{Get a size value for plotting the triad anchor titles}
\usage{
get_triad_anchor_plot_size(anchor_titles)
}
\arguments{
\item{anchor_titles}{\itemize{
\item the named list of anchor plot titles obtained from the get_triad_anchor_plot_titles function.
}}
}
\value{
An integer for the anchor plot size.
}
\description{
This might be a useful function - it is used by the workbench and plot_triad function and its return values are appropriate for Workbench sized plots.
}
