% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signifier_helpers.R
\name{get_triad_anchor_plot_titles}
\alias{get_triad_anchor_plot_titles}
\title{\itemize{
\item Get the anchor titles to plot for a triad.
}}
\usage{
get_triad_anchor_plot_titles(
  triad_id,
  display_anchor_means,
  anchor_means,
  framework_object
)
}
\arguments{
\item{triad_id}{\itemize{
\item The signifier id of the triad.
}}

\item{display_anchor_means}{\itemize{
\item If TRUE, display the anchor means on the titles.
}}

\item{anchor_means}{\itemize{
\item the anchor means to display - obtained from the calculate_triad_means function.
}}

\item{framework_object}{\itemize{
\item the sensemakerframeworkr object associated with the triad.
}}
}
\value{
A named list of the triad anchor titles.
}
\description{
\itemize{
\item Get the anchor titles to plot for a triad.
}
}
