% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{get_triad_count_canvas}
\alias{get_triad_count_canvas}
\title{Get the zone labeled triad canvas used as triad background in a ggplot x,y zone count and percentage plot.}
\usage{
get_triad_count_canvas()
}
\value{
The zone demarcated triad canvas
}
\description{
Get zone demarcated triad canvas - used when showing zone counts and percentages.
}
