% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{get_triad_standard_canvas}
\alias{get_triad_standard_canvas}
\title{Get the standard triad canvas used as triad background in a ggplot x,y plot.}
\usage{
get_triad_standard_canvas()
}
\value{
The standard triad png file object
}
\description{
Get standard triad canvas
}
