% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{get_triad_zone_canvas}
\alias{get_triad_zone_canvas}
\title{Get the zone labeled triad canvas used as triad background in a ggplot x,y plot.}
\usage{
get_triad_zone_canvas()
}
\value{
The triad zone labeled canvas
}
\description{
Get triad zone labeled canvas as a useful annotation explaining zone labels.
}
