% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signifier_helpers.R
\name{group_mean_ternary}
\alias{group_mean_ternary}
\title{Calculate the mean for the 3 columns of a ternary data matrix}
\usage{
group_mean_ternary(
  X,
  mean_type = "geometric",
  zero_logic = "remove",
  total = 1
)
}
\arguments{
\item{X}{\itemize{
\item 3 column ternary data.
}}

\item{mean_type}{\itemize{
\item default "geometric, Use the geometric mean otherwise the arithmetic.
}}

\item{zero_logic}{\itemize{
\item default "remove", values "small_value" or "remove". If "remove" all records with a zero value are removed, otherwise they are set to a value of 0.00001
}}

\item{total}{\itemize{
\item default 1, values 1 or 100, which sets the 3 return values to sum to 1 or 100.
}}
}
\value{
The confidence interval data to graph in Ternary.
}
\description{
Supply a 3 column matrix and return the means of each column
}
