% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{is_not_blank_null_na}
\alias{is_not_blank_null_na}
\title{generic is not blank, null or na helper function.}
\usage{
is_not_blank_null_na(x)
}
\arguments{
\item{x}{\itemize{
\item the string to check
}}
}
\value{
TRUE or FALSE - TRUE if the text is not NA, NULL or blank.
}
\description{
Is a string NOT blank or null or NA
}
