% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{is_valid_name}
\alias{is_valid_name}
\title{Is a string passed a valid name.}
\usage{
is_valid_name(x)
}
\arguments{
\item{x}{A vector character names.}
}
\value{
A list of logicals indicating whether each string is a name or not.
}
\description{
Determines whether a character string is a valid R name
}
