% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signifier_helpers.R
\name{mean_conf_region}
\alias{mean_conf_region}
\title{Calculate the confidence region around the meanfor a set of ternary data}
\usage{
mean_conf_region(
  X,
  level = 0.95,
  npoints = 200,
  method = c("hotelling", "normal")
)
}
\arguments{
\item{X}{\itemize{
\item 3 column ternary data.
}}

\item{level}{\itemize{
\item default 0.95, the confidence interval to calculate.
}}

\item{npoints}{\itemize{
\item default 200, the number of points to use.
}}

\item{method}{\itemize{
\item "hotelling" or "normal"
}}
}
\value{
The confidence interval data to graph in Ternary.
}
\description{
Supply a 3 column matrix and return the confidence interval around the mean
}
