% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signifier_helpers.R
\name{normalize_ternary}
\alias{normalize_ternary}
\title{normalise ternary data to sum to 1}
\usage{
normalize_ternary(X)
}
\arguments{
\item{X}{\itemize{
\item the data being normalised. 3 columns, top, right, left - the Ternary package standard.
}}
}
\value{
The input data with each row suming to 1 over the three columns.
}
\description{
Code to normalise data to sum to 1 - can also be done with compositions package but this is package neutral.
}
