% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{perform_adonis2}
\alias{perform_adonis2}
\title{Return the statistical output from a Non-Parametric PERMANOVA (Permutation-based Multivariate Analysis of Variance) test.}
\usage{
perform_adonis2(data, control_var, non_parametric = TRUE, b_value = 1000)
}
\arguments{
\item{data}{\itemize{
\item The data frame with the data columns to test.
}}

\item{control_var}{\itemize{
\item A vector of length nrows(data) with the test values.
}}

\item{non_parametric}{\itemize{
\item Default TRUE, to perform a non-parametric  test. This value must always be TRUE.
}}

\item{b_value}{The number of permutations to perform on a non-parametric test.}
}
\value{
An object of class anova.cca"  "anova"      "data.frame" with the test result.
}
\description{
Non-Parametric PERMANOVA (Permutation-based Multivariate Analysis of Variance) test
}
