% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{perform_dqdist.etest}
\alias{perform_dqdist.etest}
\title{Return the statistical output from an Energy non_parametric test.}
\usage{
perform_dqdist.etest(data, control_var, non_parametric = TRUE, b_value = 1000)
}
\arguments{
\item{data}{\itemize{
\item The data frame with the data columns to test.
}}

\item{control_var}{\itemize{
\item A vector of length nrows(data) with the test values.
}}

\item{non_parametric}{\itemize{
\item Default TRUE, to perform a non-parametric  test. This value must always be TRUE.
}}

\item{b_value}{The number of permutations to perform on a non-parametric test.}
}
\value{
An object of class "htest" with the hotelling T^2 test result.
}
\description{
Perform an Energy test on columns of data with a control variable. This is a non-parametric test.
}
