% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{perform_hotelling}
\alias{perform_hotelling}
\title{Return the statistical output from a Hotelling T test.}
\usage{
perform_hotelling(data, control_var, non_parametric = TRUE, b_value = 1000)
}
\arguments{
\item{data}{\itemize{
\item The data frame with the data columns to test.
}}

\item{control_var}{\itemize{
\item A vector of length nrows(data) with the test values.
}}

\item{non_parametric}{\itemize{
\item Default TRUE, to perform a non-parametric T^2 test, otherwise FALSE to perform a parametric test.
}}

\item{b_value}{The number of permutations to perform on a non-parametric test.}
}
\value{
An object of class "hotelling.test" with the hotelling T^2 test result.
}
\description{
Perform a hotelling T^2 test on columns of data with a control variable.
}
