% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{perform_npmv_nonpartest}
\alias{perform_npmv_nonpartest}
\title{Return the statistical output from an MANOVA test using the npmv package.}
\usage{
perform_npmv_nonpartest(
  data,
  control_var,
  non_parametric = TRUE,
  b_value = 1000
)
}
\arguments{
\item{data}{\itemize{
\item The data frame with the data columns to test.
}}

\item{control_var}{\itemize{
\item A vector of length nrows(data) with the test values.
}}

\item{non_parametric}{\itemize{
\item Default TRUE, to perform a non-parametric  test. This value must always be TRUE.
}}

\item{b_value}{The number of permutations to perform on a non-parametric test.}
}
\value{
An object of class "list" (of 2 data frames) with the MANOVA test result. The first list entry contains the results, the second the relative effects for each response variable.
}
\description{
Perform an MANOVA test on columns of data with a control variable. This is a non-parametric test.
}
