% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{plot_dyad}
\alias{plot_dyad}
\title{Plot a standard dyad graph.}
\usage{
plot_dyad(
  filtered_data,
  full_data,
  dyad_id,
  framework_object,
  dyad_title_size = "default",
  dyad_x_axis_title_size = "default",
  dyad_x_axis_title_wrap_length = 35
)
}
\arguments{
\item{filtered_data}{\itemize{
\item Must be supplied. Data frame that includes the dyad XR columns with filtered (if any) signifiers to plot.
}}

\item{full_data}{\itemize{
\item Must be supplied. Data frame that includes the dyad XR columns with all signifiers for the capture.
}}

\item{dyad_id}{\itemize{
\item Must be supplied. The dyad_id to be plotted.
}}

\item{framework_object}{\itemize{
\item Must be supplied. The framework object from the sensemakerdatar instance object. (this is sensemakerdatar::sm_framework)
}}

\item{dyad_title_size}{The font size of the dyad graph title, default value "default" let the algorithm calculate, otherwise numeric font size.}

\item{dyad_x_axis_title_size}{\itemize{
\item The font size of the dyad x axis title, default value "default" let the algorithm calculate, otherwise numeric font size.
}}

\item{dyad_x_axis_title_wrap_length}{\itemize{
\item Default 35, number of characters to wrap left and right x axis title text.
}}
}
\value{
A ggplot graph object of the dyad
}
\description{
Plot a SenseMaker® defined dyad with ggplot (note requires the sensemakerframeworkr object)
}
