% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{plot_emotions_over_time}
\alias{plot_emotions_over_time}
\title{Plot a timed line graph of the sentiment emotions of the freetext ids passed to the function.}
\usage{
plot_emotions_over_time(sentiment_filters, freetexts_to_plot, framework_data)
}
\arguments{
\item{sentiment_filters}{\itemize{
\item A data frame or name of a csv file with columns id and title. The id column contains the names of data queries and the title column the names to use in the graph print.
}}

\item{freetexts_to_plot}{\itemize{
\item default NULL, A vector of freetext signifier ids for textual data If NULL all freetext signifiers set as a fragment are plotted.
}}

\item{framework_data}{\itemize{
\item the sensemakerdatar object for the framework.
}}
}
\description{
This function plots a line graph over time of the sentiment emotions.
}
