% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{plot_list}
\alias{plot_list}
\title{Plot a list bar chart.}
\usage{
plot_list(
  filtered_data,
  full_data,
  sig_id,
  framework_object,
  bar_colour = "#7084B8",
  bar_transparency = 1,
  border_colour = "black",
  show_percentages_counts = "count",
  title_size = NULL,
  display_guide = FALSE,
  display_stats = TRUE,
  colour_sig_id = NULL,
  colour_vector = NULL,
  colour_package = "grDevices",
  colour_function = "rainbow"
)
}
\arguments{
\item{filtered_data}{\itemize{
\item Must be supplied. Data frame that includes the triad x and y columns with filtered (if any) signifiers to plot.
}}

\item{full_data}{\itemize{
\item Must be supplied. Data frame that includes the triad x and y columns with all signifiers for the capture.
}}

\item{sig_id}{\itemize{
\item Must be supplied. The triad_id to be plotted.
}}

\item{framework_object}{\itemize{
\item Must be supplied. The framework object from the sensemakerdatar instance object. (this is sensemakerdatar::sm_framework)
}}

\item{bar_colour}{\itemize{
\item default "#7084B8". The colour of the graph dots. A Character string of any valid R colour format, such as hex values or colour names.
}}

\item{bar_transparency}{\itemize{
\item default 1. The transparency (or alpha) value for the dots. A numeric value between 0 and 1.
}}

\item{border_colour}{\itemize{
\item default "black" - any valud R colour for the bar border colour.
}}

\item{show_percentages_counts}{\itemize{
\item default "counts" - options "counts", "percentages", "none" to display on the top of each bar.
}}

\item{title_size}{\itemize{
\item default NULL. The size of the title. If NULL it is calculated.
}}

\item{display_guide}{\itemize{
\item default FALSE - whether to display the colour legend if colour_sig_id set.
}}

\item{display_stats}{\itemize{
\item default TRUE - whether to display the stats data at the bottom of the graph.
}}

\item{colour_sig_id}{\itemize{
\item default NULL, don't colour by a list signifier, otherwise the signifier id of the list to colour by.
}}

\item{colour_vector}{\itemize{
\item default NULL, a vector of valid R colour codes (alpha, hex etc.) of length the number of list items in the signifier id used for the colour.
}}

\item{colour_package}{\itemize{
\item default "grDevices" - any package that contains colour palletes that can return a list of colours - this is used if colour vector is NULL.
}}

\item{colour_function}{\itemize{
\item default "rainbow", which is a default for grDevices.
}}
}
\value{
A ggplot graph object of the list (multi-choice question)
}
\description{
Plot a SenseMaker defined list (multi-choice question) with ggplot (note requires the sensemakerframeworkr object)
}
