% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{plot_tern_means_by_list_id}
\alias{plot_tern_means_by_list_id}
\title{Plot triad means for a list id.}
\usage{
plot_tern_means_by_list_id(
  filtered_data,
  triad_id,
  list_id,
  framework_object,
  list_item_ids = NULL,
  colour_vector = NULL,
  show_mean = TRUE,
  mean_colour = NULL,
  mean_shape = NULL,
  mean_zero_logic = "small_value",
  mean_size = 2,
  colour_package = "RColorBrewer",
  package_palette = "Set1",
  colour_direction = 1,
  viridis_default_start = 0,
  viridis_default_end = 0.6,
  show_colour_legend = TRUE,
  legend_title_colour = "black",
  legend_title_size = 8,
  legend_text_colour = "black",
  legend_text_size = 8,
  dot_size = 0.5,
  dot_transparency = 1,
  show_confidence_intervals = TRUE,
  confidence_value = 0.05,
  confidence_size = 2,
  title_colour = "black",
  title_size = 12,
  anchor_colour = "black",
  anchor_size = 10
)
}
\arguments{
\item{filtered_data}{\itemize{
\item The dataframe of data to print.
}}

\item{triad_id}{\itemize{
\item The triad to be plotted.
}}

\item{list_id}{\itemize{
\item The signifier id of the list to have printed.
}}

\item{framework_object}{\itemize{
\item The framework object pertaining to the collection.
}}

\item{list_item_ids}{\itemize{
\item default NULL, list item ids for the list_id. If NULL all the items are plotted.
}}

\item{colour_vector}{\itemize{
\item default NULL, if provided, a vector of the same length as df_list with the colours to plot each of the data frames passed.
}}

\item{show_mean}{\itemize{
\item default TRUE, display the mean or not of the data.
}}

\item{mean_colour}{\itemize{
\item default NULL, if provided, a vector of the same length as df_list with the colours for the means. If NULL, same colour used as in colour_vector.
}}

\item{mean_shape}{\itemize{
\item default NULL, if provided, a vector of the same length as df_list, with the shapes of the means. If NULL, then "circle" used.
}}

\item{mean_zero_logic}{\itemize{
\item default "small_value" otherwise "remove" can be specified to remove NA records.
}}

\item{mean_size}{\itemize{
\item default 2, the dot size of the mean.
}}

\item{colour_package}{\itemize{
\item default "RColorBrewer". Can also use "viridis". If the colour_vector is null then the plot function will assign values from the colour package.
}}

\item{package_palette}{\itemize{
\item default "Set1" for the RColorBrewer palette, use sensemakerhelpersr::get_RColorBrewer_pal_names() to get a list of RColorBrewer names. Viridis has "A" - "H".
}}

\item{colour_direction}{\itemize{
\item default 1. 1 for left to right colour selection from the palette, -1 for reverse.
}}

\item{viridis_default_start}{\itemize{
\item default 0, A value between 0-1 for start colour value in the selected viridis palette. Only applicable if "viridis" selected as the colour package.
}}

\item{viridis_default_end}{\itemize{
\item default 0.6. A value between 0-1 for end colour value in the selected viridis palette. Only applicable if "viridis" selected as the colour package.
}}

\item{legend_title_size}{\itemize{
\item default 8. The size of the legend title.
}}

\item{legend_text_colour}{\itemize{
\item default "black". The colour legend text A Character string of any valid R colour format, such as hex values or colour names.
}}

\item{legend_text_size}{\itemize{
\item default 8. The size of the legend text.
}}

\item{dot_size}{\itemize{
\item default 0.5, the size of the graph dots
}}

\item{dot_transparency}{\itemize{
\item default 1, the alpha (transparency value of the dits)
}}

\item{show_confidence_intervals}{\itemize{
\item default TRUE, display the confidence intervals.
}}

\item{confidence_value}{\itemize{
\item default 0.05 - the confidence interval to draw around the mean.
}}

\item{confidence_size}{\itemize{
\item default 2, the size of the confidence line
}}

\item{title_colour}{\itemize{
\item default "black", the title colour, a character string of any valid R colour format, such as hex values or colour names.
}}

\item{title_size}{\itemize{
\item default 12, the size of the graph title.
}}

\item{anchor_colour}{\itemize{
\item default "black" the colour of the anchor titles, a character string of any valid R colour format, such as hex values or colour names.
}}

\item{anchor_size}{\itemize{
\item default 10, the size of the anchor titles.
}}
}
\value{
A ggplot graph object of the triad.
}
\description{
Plot triad means of passed in data frame data and confidence intervals for comparison - requires the sensemakerframeworkr object
description
}
