% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{plot_tern_triad}
\alias{plot_tern_triad}
\title{Plot a triad graph using ggtern. Equivalent to the TernTriad tab in the Workbench.}
\usage{
plot_tern_triad(
  filtered_data,
  full_data,
  triad_id,
  framework_object,
  dot_size = 0.5,
  dot_colour = "black",
  dot_transparency = 1,
  opaque_filtered = FALSE,
  opaque_filter_dot_size = 0.4,
  opaque_filter_dot_transparency = 0.5,
  opaque_filter_dot_colour = "blue",
  anchor_label_size = "default",
  anchor_label_size_print = "default",
  show_mean = TRUE,
  mean_size = 2,
  mean_colour = "blue",
  mean_shape = "circle",
  mean_zero_logic = "small_value",
  graph_title = NULL,
  title_colour = "black",
  title_size = NULL,
  show_confidence_intervals = FALSE,
  confidence_interval_colour = "blue",
  confidence_interval_Size = 0.55,
  confidence_intervals = 0.05,
  show_variance = FALSE,
  variance_colour = "#08E5F5",
  variance_transparency = 0.3,
  show_contours = FALSE,
  countour_size = 0.5,
  contour_colour = "blue",
  show_contour_fill = FALSE,
  fill_transparency = 0.5,
  show_fill_legend = FALSE,
  fill_brewer_palette = "Spectral",
  colour_sig_id = NULL,
  colour_vector = NULL,
  colour_package = "RColorBrewer",
  package_palette = "Set1",
  colour_direction = 1,
  viridis_default_start = 0,
  viridis_default_end = 0.6,
  show_colour_legend = TRUE,
  legend_title_colour = "black",
  legend_title_size = 8,
  legend_text_colour = "black",
  legend_text_size = 8,
  return_as_print = FALSE,
  put_title_in_top = FALSE
)
}
\arguments{
\item{filtered_data}{\itemize{
\item Must be supplied. Data frame that includes the triad x and y columns with filtered (if any) signifiers to plot.
}}

\item{full_data}{\itemize{
\item Must be supplied. Data frame that includes the triad x and y columns with all signifiers for the capture.
}}

\item{triad_id}{\itemize{
\item Must be supplied. The triad_id to be plotted.
}}

\item{framework_object}{\itemize{
\item Must be supplied. The framework object from the sensemakerdatar instance object. (this is sensemakerdatar::sm_framework)
}}

\item{dot_size}{\itemize{
\item default 0.5, the size of the graph dots
}}

\item{dot_colour}{\itemize{
\item default black, the dot colour.
}}

\item{dot_transparency}{\itemize{
\item default 1, the transparency (alpha value) of the dots. Numeric between 0 and 1.
}}

\item{opaque_filtered}{\itemize{
\item default FALSE. If TRUE, the graph will display filtered out data as well as the current filtered data.
}}

\item{opaque_filter_dot_size}{\itemize{
\item default 0.4. The size of the filtered out data displayed if opaque_filtered set to TRUE.
}}

\item{opaque_filter_dot_transparency}{\itemize{
\item default 0.5. The transparency (or alpha) value for the filtered out data displayed if opaque_filtered set to TRUE. A numeric value between 0 and 1.
}}

\item{opaque_filter_dot_colour}{\itemize{
\item default "blue". The colour of the filtered out data displayed if the opaque_filtered set to TRUE. A character string of any valid R colour format, such as hex values or colour names.
}}

\item{anchor_label_size}{\itemize{
\item default "default" otherwise a numeric value to specify the size of the anchor labels. If "default" the size is calculated based on the length of the text.
}}

\item{anchor_label_size_print}{\itemize{
\item default "default" otherwise a numeric value to specify the size of the anchor labels on a print. If "default" the size is calculated based on the length of the text.
}}

\item{show_mean}{\itemize{
\item default TRUE, otherwse FALSE. If TRUE, displays the geometric mean in the ternary graph.
}}

\item{mean_size}{\itemize{
\item default 2, the dot size of the mean.
}}

\item{mean_colour}{\itemize{
\item default "blue", the colour of the mean if displayed.
}}

\item{mean_shape}{\itemize{
\item default "circle", otherwise "diamond", the shope of the mean being displayed.
}}

\item{mean_zero_logic}{\itemize{
\item default "small_value" otherwise "remove" can be specified to remove NA records. "small_value" sets zero anchors to a very small value.
}}

\item{graph_title}{\itemize{
\item default NULL, if NULL the signifier title is used for the graph title, otherwise the value passed here.
}}

\item{title_colour}{\itemize{
\item default black, The colour of the graph title. A Character string of any valid R colour format, such as hex values or colour names.
}}

\item{title_size}{\itemize{
\item default NULL, the size of the graph title.
}}

\item{show_confidence_intervals}{= default FALSE, if TRUE, show confidence intervals in the graph.}

\item{confidence_interval_colour}{\itemize{
\item default blue, a Character string of any valid R colour format, such as hex values or colour names.
}}

\item{confidence_interval_Size}{\itemize{
\item default 0.55, the size of the confidence interval lines.
}}

\item{confidence_intervals}{\itemize{
\item default 0.05, a vector of confidence intervals to display with each value between 0 and 1.
}}

\item{show_variance}{\itemize{
\item default FALSE, if TRUE, show the variance around the mean as a shaded area.
}}

\item{variance_colour}{\itemize{
\item default hex "#08E5F5",  a character string of any valid R colour format, such as hex values or colour names.
}}

\item{variance_transparency}{\itemize{
\item default 0.3, the transparency (alpha value) of the variance shaded area. A value between 0 and 1.
}}

\item{show_contours}{\itemize{
\item default FALSE, if TRUE show contour lines based on kernel smoothing function.
}}

\item{countour_size}{\itemize{
\item default 0.5, the size of the contour lines.
}}

\item{contour_colour}{\itemize{
\item defaykt "blue", the colour of the contour lines, a character string of any valid R colour format, such as hex values or colour names.
}}

\item{show_contour_fill}{\itemize{
\item default FALSE, show heat map, or filled contour based on kernel smoothing function.
}}

\item{fill_transparency}{\itemize{
\item default 0.5, the transparency (alpha value) of the contour fill regions.
}}

\item{show_fill_legend}{\itemize{
\item default FALSE, if TRUE, show the legend with the probability values for the contour/fill levels.
}}

\item{fill_brewer_palette}{\itemize{
\item default "Spectral", the RColorBrewer palette used to colour the contour fills. Use sensemakerhelpersr::get_colour_palette_names() to see all the palette names available.
}}

\item{colour_sig_id}{\itemize{
\item default NULL, A single select list (multi-choice question) id to use as a colour aesthetic in the graph.
}}

\item{colour_vector}{\itemize{
\item default NULL, a vector of length the number of list items providing the colours to use for each item in the colour_sig_id colouring.
}}

\item{colour_package}{\itemize{
\item default "RColorBrewer". Can also use "viridis". If the colour_vector is null then the plot function will assign values from the colour package.
}}

\item{package_palette}{\itemize{
\item default "Set1" for the RColorBrewer Set1 palette. values of "Dark2", "Set1", "Set2" or "Set3" recommended but others can also work. Viridis has "A" - "H".
}}

\item{colour_direction}{\itemize{
\item default 1. 1 for left to right colour selection from the palette, -1 for reverse.
}}

\item{viridis_default_start}{\itemize{
\item default 0, A value between 0-1 for start colour value in the selected viridis palette. Only applicable if "viridis" selected as the colour package.
}}

\item{viridis_default_end}{\itemize{
\item default 0.6. A value between 0-1 for end colour value in the selected viridis palette. Only applicable if "viridis" selected as the colour package.
}}

\item{show_colour_legend}{\itemize{
\item default TRUE, show the colour legend.
}}

\item{legend_title_colour}{\itemize{
\item default "black", the colour of the colour legend title, a character string of any valid R colour format, such as hex values or colour names.
}}

\item{legend_title_size}{\itemize{
\item default 8. The size of the colour legend title.
}}

\item{legend_text_colour}{\itemize{
\item default "black". The colour legend text A Character string of any valid R colour format, such as hex values or colour names.
}}

\item{legend_text_size}{\itemize{
\item default 8. The size of the legend text.
}}

\item{return_as_print}{\itemize{
\item default FALSE, returns the ggtern object otherwise a print of the object.
}}

\item{put_title_in_top}{\itemize{
\item default FALSE, if TRUE will place the graph title 2 lines above the triad top. This will reduce the space between the title and the top label.
}}
}
\value{
A ggtern graph object of the triad either as a plot object or printed object (depending on the return_as_print parameter value).
}
\description{
Plot triad using ggtern
}
