% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{produce_keyness_pair_graphs}
\alias{produce_keyness_pair_graphs}
\title{Plot keyness bar charts for pairs of data queries.}
\usage{
produce_keyness_pair_graphs(
  keyness_pairs,
  framework_data,
  freetext_ids = NULL,
  filters = NULL,
  filter_names = NULL,
  data_to_use = NULL,
  source_field = "source",
  check_enough_rows = FALSE,
  min_term_freq = 3,
  languages = "en",
  stem_text = TRUE
)
}
\arguments{
\item{keyness_pairs}{\itemize{
\item A data frame with columns "from_id", "to_id", "from_colour" and "to_colour". The from and to ids are data queries stored in the data list (public field). The colours are valid R colour names or codes.
}}

\item{framework_data}{\itemize{
\item The framework sensemakerdatar object.
}}

\item{freetext_ids}{\itemize{
\item default NULL, A vector of freetext signifier ids for textual data If NULL all freetext signifiers set as a fragment are plotted.
}}

\item{filters}{\itemize{
\item named list containing the filters to additionally apply to the appended dataset.
}}

\item{filter_names}{\itemize{
\item names for the filters - normally the names R function on the filters.
}}

\item{data_to_use}{\itemize{
\item Default NULL, as the data will be the combined pairs from the keyness pairs, but for special data processing using filters, the data data.frame is likely to be a specially constructed one.
}}

\item{source_field}{\itemize{
\item Default "source", this is for the corpus creation for separating out the datasets. Normally "source" created in the sensemakerdatar pair dataset creation but when using filters with specific datasets, this could be something different.
}}

\item{check_enough_rows}{\itemize{
\item Default FALSE, when using very small data sets especially with small amounts of text and little data records, further checks on the corpus is needed. This is a very slow algorithm so use with care.
}}

\item{min_term_freq}{\itemize{
\item Default 3, number of occurrence of a term before it it is accepted into the corpus.
}}

\item{languages}{\itemize{
\item Default "en", a vector of supported 2 character language codes for use in stop words and stemming.
}}

\item{stem_text}{\itemize{
\item Default TRUE, whether to stem the text.
}}
}
\description{
This function plots a keyness indicator bar chart graph for pairs of data queries.
}
