% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{produce_tern_pair_means_graphs_by_triad}
\alias{produce_tern_pair_means_graphs_by_triad}
\title{Plot tern pair means based on data queries.}
\usage{
produce_tern_pair_means_graphs_by_triad(
  tern_pairs,
  framework_data,
  triads_to_plot = NULL
)
}
\arguments{
\item{tern_pairs}{\itemize{
\item A data frame with columns "from_id", "to_id", "from_colour" and "to_colour". The from and to ids are data queries stored in the data list (public field). The colours are valid R colour names or codes.
}}

\item{framework_data}{\itemize{
\item The framework sensemakerdata object.
}}

\item{triads_to_plot}{\itemize{
\item default NULL, A vector of triad ids to plot. If NULL all keep_only_include triads are plotted.
}}
}
\description{
This function plots the means and 5\% confidence intervals for a list of triads.
}
