% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{removeHTML}
\alias{removeHTML}
\title{generic remove html helper function}
\usage{
removeHTML(input_string)
}
\arguments{
\item{input_string}{\itemize{
\item the string to remove html tags.
}}
}
\value{
The string with any html removed.
}
\description{
Remove html string from an input string
}
