% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{return_chisq_test}
\alias{return_chisq_test}
\title{Perform and return a chi-square test directly from the count table.}
\usage{
return_chisq_test(
  dataTabMatrix,
  row_names,
  col_names = NULL,
  round_digits = 0,
  residual_threshold = 4,
  p_threshold = 0.05
)
}
\arguments{
\item{dataTabMatrix}{\itemize{
\item The data in table form, can be matrix, tibble, dataframe but must contain only count columns.
}}

\item{row_names}{\itemize{
\item The names of the rows.
}}

\item{col_names}{\itemize{
\item The names of the columns. If NULL it is assumed that the columns are alraedy named as named columns can be used in the chi-squared test.
}}

\item{round_digits}{\itemize{
\item default 0, z, z^2, expected value table rounding digits.
}}

\item{residual_threshold}{\itemize{
\item default 4, the threshold for residual calculation on residual squared values.
}}

\item{p_threshold}{\itemize{
\item default 0.05, the p-value threshold for null hypothesis test.
}}
}
\value{
Returns a named list.containing the residual calculations. z, zsqr  the data count matrix, expected value matrix, p_value, test_result accept null hypothesis TRUE or FALSE, .
}
\description{
This function will perform goodness of fit tests and return the results for a populated count table.
}
