% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensemakerhelpersr-package.R
\docType{package}
\name{sensemakerhelpersr-package}
\alias{sensemakerhelpersr}
\alias{sensemakerhelpersr-package}
\title{sensemakerhelpersr: generic helpers for working with SenseMaker® data}
\description{
This package provides helper functions for the sensemakerframeworkr and sensemakerdatar packages and for programmers working with SenseMaker® data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PeterStanbridge/sensemakerhelpersr}
  \item Report bugs at \url{https://github.com/PeterStanbridge/sensemakerhelpersr/issues}
}

}
\author{
\strong{Maintainer}: Peter Stanbridge \email{peter.stanbridge@cognitive-edge.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
