% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signifier_helpers.R
\name{ternary_to_xy}
\alias{ternary_to_xy}
\title{ternary to <x,y>}
\usage{
ternary_to_xy(X)
}
\arguments{
\item{X}{\itemize{
\item the data being converted - 3 columns, top, right, left - the Ternary package standard.
}}
}
\value{
The data in two columns, x and y.
}
\description{
Take a 3 column ternary data and calculate the x,y values to return for standard plot.
}
