% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{word_frequency_plot}
\alias{word_frequency_plot}
\title{Plot a frequency graph of the words used between two different queries.}
\usage{
word_frequency_plot(
  frequency_graph_pairs,
  freetexts_to_plot,
  data_object,
  use_stem = FALSE
)
}
\arguments{
\item{frequency_graph_pairs}{\itemize{
\item A data frame or name of a csv file with columns "from_id" and "to_id" containing data query names.
}}

\item{freetexts_to_plot}{\itemize{
\item default NULL, A vector of freetext signifier ids for textual data If NULL all freetext signifiers set as a fragment are plotted.
}}

\item{data_object}{\itemize{
\item the sensemakerdatar object for the framework.
}}

\item{use_stem}{\itemize{
\item default NULL, If TRUE, the graph will accumulate stemmed word usage.
}}
}
\description{
Plot a frequency graph of the words used between two different queries.
}
