% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{wrap_text}
\alias{wrap_text}
\title{wrap a text string at a given length}
\usage{
wrap_text(text_string, wrap_length = 30, replacement_value = "")
}
\arguments{
\item{text_string}{\itemize{
\item the string to be wrapped.
}}

\item{wrap_length}{\itemize{
\item Default 30, the position length for the text to be wrapped.
}}

\item{replacement_value}{\itemize{
\item A string to use as a replacement in the wrap break - i.e. instead of breaking.
}}
}
\value{
The wrapped text_string
}
\description{
wrap a text string at a given length
}
