% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signifier_helpers.R
\name{xy_to_ternary}
\alias{xy_to_ternary}
\title{<x,y> to ternary}
\usage{
xy_to_ternary(xy)
}
\arguments{
\item{X}{\itemize{
\item the data being converted - 2 columns, x and y values.
}}
}
\value{
The data in three columns, top, right, left - the Ternary package standard.
}
\description{
Take a 2 column ternary based x,y data set and calculate the three ternary values.
}
