% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signifier_helpers.R
\name{calculate_triad_means}
\alias{calculate_triad_means}
\title{Calculate triad means}
\usage{
calculate_triad_means(
  data,
  triad_id,
  mean_type,
  framework_object,
  zero_logic = "small_value",
  for_ggtern = FALSE
)
}
\arguments{
\item{data}{\itemize{
\item The data containing the triad top, left and right columns. This can be the full data dataframe, provided the triad's 3 columns are present.
}}

\item{triad_id}{\itemize{
\item The signifier id of the triad to calculate means.
}}

\item{mean_type}{\itemize{
\item either "geometric" or "arithmetic".
}}

\item{zero_logic}{\itemize{
\item either "small_value" or "remove". Small value means to take a zero value and give it a very small replacement. Otherwise ignore this fragment.
}}

\item{for_ggtern}{\itemize{
\item Default FALSE, if true, the returned dataframe containing the mean is in a format ready to plot in ggtern.
}}
}
\value{
The triad means as 3 values.
}
\description{
Calculate the triad means using specified approach
}
