% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signifier_helpers.R
\name{calculate_triad_zone}
\alias{calculate_triad_zone}
\title{get a triad value's zone}
\usage{
calculate_triad_zone(top, left, right)
}
\arguments{
\item{top}{percentage triad top value (between 0 and 100).}

\item{left}{percentage triad left value (between 0 and 100).}

\item{right}{percentage triad right value (between 0 and 100).}
}
\value{
A single character returning the triad zone
}
\description{
This function returns the zone value for a triad based on passing in a left, top and right compositional value.
}
\details{
get a triad value's zone
}
\examples{
t <- 30
l <- 65
r <- 5
calculate_triad_zone(t, l, r)
}
