% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{get_correlations_by_type}
\alias{get_correlations_by_type}
\title{Get the correlation data for signifiers of a from and to signifier type}
\usage{
get_correlations_by_type(df, fw, from_type, to_type, round_digits = 0)
}
\arguments{
\item{df}{\itemize{
\item The data to perform the tests - this is a data frame containing at least all the columns needed for the signifier types passed in. .
}}

\item{fw}{\itemize{
\item The framework definition object.
}}

\item{from_type}{\itemize{
\item Either "list" or one of the shape types ("dyad", "triad", "stones").
}}

\item{to_type}{\itemize{
\item Either "list" or one of the shape types ("dyad", "triad", "stones").
}}

\item{round_digits}{\itemize{
\item default 0, z, z^2, expected value table rounding digits.
}}
}
\value{
Returns a named list. ids_to_output contains the dataframe of signifier id pairs included in the output. sig_residuals a list of length the number of correlation pairs, It contains the "data", "expected", "residuals", "residuals_sqr", "p-value".
}
\description{
This function will perform goodness of fit tests and return the results for each signifier combination of the from and to signifier types. For lists, these will be only the single select signifier mcqs and the date columns.
The actual tests are performed on the zone and region categorical columns for triads, dyads and stones.
}
