% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_helpers.R
\name{get_residuals}
\alias{get_residuals}
\title{Get the correlation data for signifiers of a from and to signifier type}
\usage{
get_residuals(df, fw, from_col, to_col, round_digits = 0)
}
\arguments{
\item{df}{\itemize{
\item The data to perform the tests - this is a data frame containing at least all the columns needed for the signifier types passed in. .
}}

\item{fw}{\itemize{
\item The framework definition object.
}}

\item{from_col}{\itemize{
\item The from data column name in data frame df for correlation calculation
}}

\item{to_col}{\itemize{
\item The to data column name in data frame df for correlation calculation. Must not be the same as from_col
}}

\item{round_digits}{\itemize{
\item default 0, z, z^2, expected value table rounding digits.
}}
}
\value{
Returns a named list.containing the residual calculations. z, zsqr  the data count matrix, expected value matrix, p_value and test_result accept null hypothesis TRUE or FALSE.
}
\description{
This function will perform goodness of fit tests and return the results for each signifier combination of the from and to signifier types. For lists, these will be only the single select signifier mcqs and the date columns.
The actual tests are performed on the zone and region categorical columns for triads, dyads and stones.
}
