% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{produce_keyness_pair_graphs}
\alias{produce_keyness_pair_graphs}
\title{Plot keyness bar charts for pairs of data queries.}
\usage{
produce_keyness_pair_graphs(
  keyness_pairs,
  framework_data,
  freetext_ids = NULL,
  languages = "en"
)
}
\arguments{
\item{keyness_pairs}{\itemize{
\item A data frame with columns "from_id", "to_id", "from_colour" and "to_colour". The from and to ids are data queries stored in the data list (public field). The colours are valid R colour names or codes.
}}

\item{framework_data}{\itemize{
\item The framework sensemakerdatar object.
}}

\item{freetext_ids}{\itemize{
\item default NULL, A vector of freetext signifier ids for textual data If NULL all freetext signifiers set as a fragment are plotted.
}}
}
\description{
This function plots a keyness indicator bar chart graph for pairs of data queries.
}
